SelectiveMAE/requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aenum==3.1.15
affine==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.3.2
causal-conv1d @ file:///home/wangfengxiang/remote-sensing/mamba/Crossmae/causal-conv1d/causal_conv1d-1.4.0%2Bcu118torch2.0cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=814b3c7b13316011192630f9ee9fc17a3a129c4480c8c5eb8812b96bd242545b
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmake==3.25.0
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
filelock==3.9.0
fiona==1.9.6
flash-attn @ file:///home/wangfengxiang/remote-sensing/other/Crossmae/flash_attn-2.5.1.post1%2Bcu118torch2.0cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=7e06e362647bf9345def7e6b63a5115f0a6b3a622d94ed66f7fca102b0df5a05
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.0
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.4
imagededup==0.3.2
ImageHash==4.3.1
imageio==2.33.1
importlib_resources==6.4.3
Jinja2==3.1.2
joblib==1.3.2
jsonargparse==4.32.0
kaleido==0.2.1
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.3
lightly-utils==0.0.2
lightning-utilities==0.11.6
lit==15.0.7
mamba-ssm @ file:///home/wangfengxiang/remote-sensing/mamba/Crossmae/mamba/mamba_ssm-2.2.2%2Bcu118torch2.0cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=f2cd537a0bc57ef573b6d4a87e547afa661902ebc6fb6dbbb7c6ee9a60396b2b
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.2
peft==0.3.0
pillow==10.2.0
platformdirs==4.2.2
plotly==5.23.0
# Editable install with no version control (pretrain==0.1)
-e /home/wangfengxiang/remote-sensing/other/Rmae
protobuf==4.23.4
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
rasterio==1.3.10
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
Rtree==1.3.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
sympy==1.12
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==0.9.7
tokenizers==0.15.2
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.66.1
transformers==4.38.2
triton==3.0.0
typeshed_client==2.7.0
typing_extensions==4.8.0
tzdata==2024.1
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.13
wandb==0.17.7
Werkzeug==3.0.1
XMind==1.2.0
yarl==1.9.4
